# repository: shard rules by namespace