@joestringer 2023-08-03
SGTM. Also when we've narrowed down on the user value it'll be worth updating the release note to say something about improving policy processing performance.