@joestringer 2023-11-21
Hey @odinuge, was this waiting on further feedback or discussion? Or you just haven't had time to revisit? I think that in principle this can improve Cilium quite a bit, so it would be neat to see it move forward.