@odinuge 2024-05-17
Hey @squeed! Happy to push this myself. However we need a proper solution to #30338 first, and we need _a lot_ of tests for that to avoid security regressions. If you want to start with that first that would be very appreciated!