@odinuge 2024-05-17
The easy alternative is reverting https://github.com/cilium/cilium/commit/4cb094060d5a6b93313e222516c644ef5c51da07, but not too sure how happy people would be with that. Imo. we should not introduce more "indexes" here until we are 100% the current state handles all the edge cases, with tests ensuring that is the case.