@squeed 2024-05-17
> we should not introduce more "indexes"

This seems like a good chance to convert the PolicyRepository to StateDB over the coming months.