@squeed 2024-05-24
I filed #32703, which was inspired by this but takes a different tack.