@squeed 2024-05-31
https://github.com/cilium/cilium/pull/32703 should have removed this bottleneck, I believe this can be closed, @odinuge.